requirements.txt
numpy
PyOpenGL
opencv-python
pygame